Replace `docker-compose` with `docker compose` in `bin/setup` script
The existing `bin/setup` script used the command `docker-compose` to run a rake task and fetch a secret key. This change updates the script to use the new syntax `docker compose` instead. The `docker-compose` command is deprecated in favor of the more streamlined `docker compose` command. The change ensures compatibility with the latest version of Docker and improves the codebase's adherence to recommended practices.